refactor: remove unused code in selfservice controller
The removed code is nolonger in use since linotp 3.0 or 2.90, where we
changed how url routes are matched.
`g.audit["action"]` is now "/selfservice-legacy/" when using the
selfservice controller. Even before making it the legacy controller,
it probably not reported as "selfservice/index" (which is what we are
matching in the removed code.

Therefore, the code is guaranteed to not do anything anymore and we
decided that we do not need audit logging for a selfservice access
because the only relevant info is userservice API access that is unaffected.