fix: for forward token to support push token
the push token api is used via accept or reject api endpoint. These
endpoints convert the given parameter into a dict which is then check by
the check_otp.

Within the scope of the forward token the heuristic uses the check_otp
api though to identify if the target token could answer the request
directly. For the push token the check_otp api thus woud receive a
simple string value and not a dict as expected and raises an exception
which is not handled in the forward token processing.

The fix thus handles the special case for a push token and will only
forward to the push token check_otp if the passw is really a dictionary.