fix: Don't use `LocalResponseProxy` for cookies.
The `LocalResponseProxy` is not thread-local, allowing cookies to
accumulate across different threads. Additionally, `LocalResponseProxy`
uses a list for the cookies, which means that multiple cookies with the
same name could be sent to clients. There are no definite rules to deal
with this, either on the client side or on the server side in case they
are sent back to the server.

This can result in "cookie confusion," where a client selects a session
cookie from a set generated by nearly simultaneous login attempts from
different users, assuming the identity of the user associated with that
session cookie. Because LocalResponseProxy's cookie list is emptied
after each request, issues arise only when several users attempt to log
in nearly simultaneously. An expired session can not be recalled. 

We address this by (a) moving the cookies to Flask's `g` object, which
is specifically intended to hold data that is local to a single HTTP
request/reply cycle, and (b) replacing the list with a dictionary so
every cookie name can exist only once per response.