fix: Replace `self.authUser` with `g.authUser` in `UserserviceController`.
`self.authUser`, an attribute of the `UserserviceController` class,
poses risks as it is shared across all requests. Replacing it with
`g.authUser` ensures request-specific data like `authUser` remains local
to the request, addressing potential overwrites in shared objects.

The usage of `self.authUser` can be traced back to the usage in Pylons
and was not ported correctly during the migration to Flask.

(The same problem applies to other controller object attributes and are
fixed in the next commit.)