fix: Get rid of `request_context["AuthUser"]` in `UserserviceController` (and elsewhere).
`request_context["AuthUser"]` was used to communicate the identity of
the authenticated user associated with a request to the `__after__()`
method for use in auditing/reporting. The problem this solved is that
`__after__()` is a static method and therefore did not have access to
`self.authUser`.

`g.authUser` is a perfectly suitable replacement for
`request_context["AuthUser"]`.

Note that the erroneous index `authUser` occurred in a few places.

Also note that we're fixing a few occurrences of
`request_context["AuthUser"]` in other places, too.