fix: The LinOTP restart process in Debian postinst.
It used to be possible to disable LinOTP by removing the LinOTP “site”
from Apache, because LinOTP used the Apache mod_wsgi extension to run
“inside” Apache. Nowadays, we use the `uwsgi` application server to
run LinOTP, which means that whether LinOTP runs has nothing to do
with Apache anymore.

We use systemd to control LinOTP now, which is why if LinOTP is
upgraded, we need to stop the `linotp-uwsgi.service` so systemd can
start it again (using socket activation via `linotp-uwsgi.socket`) when
another request comes in.