refactor: Remove `CONFIG_DIR` from new-style container `Dockerfile`.
In a container context, we want to encourage people to use environment
variables to configure LinOTP. This means we don't expend special
effort to support `linotp.cfg`-style configuration, and a separate
`CONFIG_DIR` to hold LinOTP configuration files is generally not
necessary. Hence it is removed from the `Dockerfile`.

It is still perfectly possible for people to mount their own directory
on `/config` (or whatever) and set `LINOTP_CFG` such that a
configuration file in that directory will be picked up. This may be
useful in migration efforts or if people have something in their
`LINOTP_CFG` file that cannot be expressed using environment
variables (remember that LinOTP's configuration file is processed as
Python code), but it should be the exception rather than the rule.

Note that we deliberately keep most of the `LINOTP_CFG` handling code
in `entrypoint.sh`, even though in the standard case it will only
output a message to the effect that `LINOTP_CFG` has not been
set. This code will keep working even with a configuration
file/directory that has been bind-mounted, and may come in useful for
debugging, because it makes clear where configuration data comes
from. (We do remove the code that, in the absence of an explicit
definition of `LINOTP_CFG` in the environment, tries to use the first
file in `CONFIG_DIR` as the main configuration file, because
`CONFIG_DIR` no longer exists.)

In any case, configuration settings in environment variables will
override configuration settings in the configuration file.