chore: Add `doas` to the main Docker image.
The Docker container runs as (non-root) `$LINOTP_USER`, but we must be
able to execute `update-ca-certificates` as root in order to enable
non-standard root CA certificates. Rather than eschewing `USER
$LINOTP_USER` in the `Dockerfile` and relying on `gunicorn` to set the
correct user (which would suck when executing LinOTP CLI commands from
outside `gunicorn`), we use `doas` – which is basically a pared-down,
streamlined version of `sudo` that should be less susceptible to
security issues – to allow the `$LINOTP_USER` to run
`update-ca-certificates` with root permissions.

We also install `ca-certificates` just to be sure that
`update-ca-certificates` is actually available.