chore: Add custom CA certificate installation mechanism.
This makes sure that any additional root-level CA certificates (whose
names must match `/usr/local/share/ca-certificates/*.crt`) are
integrated into the system trust store, by invoking
`update-ca-certificates`.

If the value of the `TLS_CA_CERTS` environment variable is not empty,
it is split into individual certificates in
`/usr/local/share/ca-certificates` before `update-ca-certificates` is
called. (We are aware that this environment variable used to be called
`SSL_TRUSTED_CERT`, but (a) we don't believe in dead-naming TLS,
and (b) *our* version supports several certificates, so there.)

It is not a problem if there are in fact no additional root-level CA
certificates; this just means that the container will take a little
bit longer to come up.