fix: wrongly aligned resize bar on Manage UI flexigrid tables on zoom
When the browser is zoomed in, the column width calculation of flexigrid breaks because it can only handle integer values for something like border width. The issue is not present when zoom is set to 100%.

In other zoom settings, borderLeftWidth can report something like `0.909091px`. The current implementation parses this to `0px` with the following brilliant code: `parseInt($(cdcol).css('borderLeftWidth'), 10))`.

To fix the scaling issues, we drop all this manual caculation things (maths is hard after all) and let jquery just report the "actual" width using ".outerWidth()".