feature: do not stop db migration if audit table could not be truncated
In case of a container environment the database priviledges are limited
to a minimum and thus the normal user is not allowed to DROP (mysql) or
TRUNCATE (postgres) database tables.
To support the linotp database migration we thus we log the error and do not
raise the SQL OperationalError and continue instead.
In case that this would not have been an priviledge problem, the error
will be raised in other places as well.