fix: linotp not starting against mysql on windows
mysql on windows uses lower case table names. This is usually no
big deal as normal sql queries are correctly quoted and parsed.

The only issue is when we parse the result of the list of table
names. Each occurence is now converting to lower case for a case-
insensitive comparison.