fix: Rewrite backslash-n to newline in the value of `TLS_CA_CERTS`.
If `TLS_CA_CERTS` comes from a Podman/Docker `--env-file` file, then
it can't have newlines, which interferes with the PEM
format. Therefore we allow the value of `TLS_CA_CERTS` to contain the
character sequence “backslash followed by lowercase n”, which we
replace with a newline when the certificates in `TLS_CA_CERTS` are
written to PEM-formatted files.