test: deleting providers
All the expected behavior when it comes to deleting providers is
presented in this test. The comments describe the protocol.
The main points are: 1-there should be no interference of policies
from other providers(of same type of course) 2-the default provider
shall only be allowed to be deleted when it is the last one left on
the list.