refactor: merge request_context creation in before_request to single location
setup_env and create_context created and filled the flask.g.request_context. Yes, they do this with very different routes, but at the end, they belong together. This will also help to untangle this mess even more when we sort out where the request-local stuff is stored in the future.