feature: Change `lib.util.get_client()` to honour `GET_CLIENT_ADDRESS_FROM_POST_DATA`.
The `get_client()` function now considers the `client=` parameter in
the HTTP post request only if the `GET_CLIENT_ADDRESS_FROM_POST_DATA`
configuration item is set to `True` (it defaults to `False`).

A deprecation warning is issued if the `client=` parameter is
used. Also, if the result of `get_client()` is not a dotted quad, we
raise a `ValueError` exception.