feature: Disable “Authorization” entry field on “System Config” dialog in LinOTP management UI.
When `GET_CLIENT_ADDRESS_FROM_POST_DATA` is `True`, it is no longer
possible to configure the IP addresses of clients who are allowed to
use the `client=` HTTP POST parameter to change their effective IP
address. This change disables the corresponding entry field on the
“System Config” dialog in the LinOTP management UI when the `client=`
parameter is ineffective. Instead, an explanatory note is displayed.