fix: failing challenge signature verification when using mysql 8 database
at least on mysql version 8, possibly also on other flavours or versions, the
column "timestamp" is rounded, which results in a different timestamp written
to database compared to the timestamp that goes into the signature that we
create before writing the challenge to database.

Some databases seem to write milliseconds, some cut off milliseconds, others
round up on high decimals which results in a one second higher timestamp
compared to the timestamp in the signature.

This closes LINOTP-2269