feature: `WORKER_THREADS=auto` scales according to number of CPU cores.
A rule of thumb for I/O-bound applications like LinOTP is to use twice
the number of CPU cores plus 1 as the number of threads in
Gunicorn. This is what we now assume as the default if
`WORKER_THREADS` is unset or set to `auto`. You can still set
`WORKER_THREADS` to an integer to use that many threads.