feature: On `/userservice/enroll`, not providing an mOTP secret creates one at random.
So far, enrolling an mOTP token required the user to provide an 8-bit
shared secret in the form of a 16-digit hexadecimal number. For
consistency with other forms of HMAC token, we want LinOTP to be able
to create a random shared secret and communicate this to a suitable
authenticator app by way of an `otpauth://` URL.

This commit lets LinOTP create a random secret for an mOTP token if
the `/userservice/enroll` API endpoint is called with `motp` as the
token type but without an `otpkey` parameter.

Note that we deliberately take the straightforward approach of using
`secrets.token_hex()` to generate the mOTP secret, rather than rely on
the rococo random-secret mechanism from the vast depths of LinOTP
which does it all from scratch. This is an example that other token
types should follow in the interest of simplicity, maintainability,
and correctness.