feature: Make `/userservice/enroll` return an enrollment URL for mOTP tokens.
This change puts mOTP tokens on the same footing as other HMAC tokens
such as TOTP, by returning an `otpauth://` URL which the UI can then
present to authenticator apps as a QR code.

It suffices to put `motp` in the list of valid token types for
`linotp.lib.apps.google_authenticator_url()`, except that for mOTP
tokens, the “seed” is a hexadecimal number which the call to
`unhexlify()` in `google_authenticator_url()` will turn into a
sequence of binary bytes, invalidating the resulting URL because the
authenticator apps do expect the hexadecimal number. Therefore we need
to make a special exception for `motp` tokens and just turn the seed
into a byte string as it is, without unhexlifying it.