test: Fix entrypoint tests.
The entrypoint tests need to be able to run the external
wait-for-the-database script, or else the tests will crash
prematurely.

Note that this means we need to extend the “implementation”
of LinOTP used in the tests to be able to deal with the
`linotp config show --values DATABASE_URI` command, because
the entrypoint requires this to work.