fix: Tweak FIDO U2F token handling for FIDO2 API compatibility.
Since FIDO U2F is no longer directly supported in web browsers, we
need to use the FIDO2 API to access FIDO U2F authenticators. This
means that FIDO U2F data coming in to LinOTP can have a tinge of
FIDO2, and we need to tweak how it is being handled in order to keep
things working. In particular this means:

  - In a FIDO2-based authentication assertion, the operation type
    field in the FIDO U2F `clientData` may be called `type` instead
    of `typ`. It is difficult to tell whether `typ` is just a typo
    that was never noticed, or whether it is required for other
    reasons, hence out of an abundance of caution we're supporting
    `type` as an alternative to `typ`, not as a replacement.

  - If an authentication assertion comes from the FIDO2 API, the
    type is `webauthn.get` rather than `navigator.id.getAssertion`,
    and we need to take this into account.

  - We need to make sure any Base64 padding at the end of the
    challenges is removed before comparing them, because the
    challenge as returned via `clientData` may or may not include
    padding while the challenge on record in LinOTP definitely
    does.