fix: Avoid problems with FIDO U2F token registration when otppin=1.
When `otppin=1`, the `check_pin()` function which is called during
FIDO U2F token registration requires a nontrivial `user` parameter
which the FIDO U2F registration code does not condescend to
provide. This makes FIDO U2F token registration fail.

The apparent solution is to detect this particular case and fall back
to `otppin=0`, which appears to work.