refactor: drop direct crypt module usage
in preparation for python 3.13 support which no-longer has the crypt
library directly. We now rely on passlib to verify these passwords
too. This does not directly grant us 3.13 compatibility but we only
have one issue to work with, and maybe libpass is matured enough
when we actually need to switch.